+++ SensateIoT.SmartEnergy.Dsmr.Processor.Data/Models/SensorMapping.cs
@@ -1,0 +1,10 @@
+﻿namespace SensateIoT.SmartEnergy.Dsmr.Processor.Data.Models
+{
+	public class SensorMapping
+	{
+		public int Id { get; set; }
+		public string PowerSensorId { get; set; }
+		public string GasSensorId { get; set; }
+		public string EnvironmentSensorId { get; set; }
+	}
+}