+++ SensateIoT.SmartEnergy.Dsmr.Processor.DataAccess/Repositories/SensorMappingRepository.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Data.SqlClient;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -10,21 +11,24 @@ namespace SensateIoT.SmartEnergy.Dsmr.Processor.DataAccess.Repositories
 {
 	public class SensorMappingRepository : AbstractRepository, ISensorMappingRepository
 	{
-		private const string SelectSensorMappings = "DsmrProcessor_SelectSensorMapping";
-		private const string SelectAllSensorMappings = "DsmrProcessor_SelectAllSensorMappings";
+		private const string SelectAllSensorMappings = "DsmrProcessor_GetDevices";
 
 		public SensorMappingRepository(string connectionString) : base(new SqlConnection(connectionString))
 		{
 		}
 
-		public Task<SensorMapping> GetSensorMapping(string sensorId, CancellationToken ct)
+		public async Task<IEnumerable<SensorMapping>> GetAllSensorsAsync(string serviceName, CancellationToken ct = default)
 		{
-			return this.QuerySingleAsync<SensorMapping>(SelectSensorMappings, "@sensorId", sensorId);
-		}
+			var result = await this.QueryAsync<Data.Models.SensorMapping>(SelectAllSensorMappings, "@processorServiceName", serviceName)
+				.ConfigureAwait(false);
 
-		public async Task<IEnumerable<SensorMapping>> GetAllSensorsAsync(CancellationToken ct = default)
-		{
-			return await this.QueryAsync<SensorMapping>(SelectAllSensorMappings).ConfigureAwait(false);
+			return result.Select(x => new SensorMapping {
+				EnvironmentSensorId = x.EnvironmentSensorId,
+				GasSensorId = x.GasSensorId,
+				Id = x.Id,
+				PowerSensorId = x.PowerSensorId,
+				LastProcessed = default
+			});
 		}
 	}
 }