+++ SensateIoT.SmartEnergy.Dsmr.Processor.Service/Application/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.ServiceProcess;
+
 using log4net;
+
 using SensateIoT.SmartEnergy.Dsmr.Processor.DataAccess.Mappers;
 using SensateIoT.SmartEnergy.Dsmr.Processor.Service.Services;