+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Product database integration
+- Sensor mapping model/dto
+
+### Updated
+- Sensor mapping data layer
+
 ## [1.2.0] - 06-04-2021
 ### Added
 - Implement data platform data deletion